+++ ariadne_codegen/client_generators/input_types.py
@@ -135,7 +135,7 @@ class InputTypesGenerator:
                 target=name,
                 annotation=annotation,
                 value=parse_input_field_default_value(
-                    node=field.ast_node, field_type=field_type
+                    node=field.ast_node, annotation=annotation, field_type=field_type
                 ),
                 lineno=lineno,
             )